performance improvement? who knows at this point
what is performance anyways? why do people make a big deal out of it? is it actually important? i assume it is, because people keep talking about it. but every time i seek out performance i only find pain. pain and suffering that is unimaginable. stuck in a prison of my own making for days just to find that i did not achieve performance. why must i suffer when i look outside and the rest of life is moving along without me? the ducks in the sky that fly past my house do not know the meaning of performance and even if they did, they would not care for it. the plant that grows in my backyard cannot even have thoughts. how blissful it must be, to not know the meaning of performance.